faiss-app/requirements.txt
google-cloud-storage
pydantic
numpy